+++ sql/ha_partition.cc
@@ -2427,7 +2427,7 @@ repeat:
   do
   {
     DBUG_PRINT("info", ("external_lock(thd, %d) iteration %d",
-                        lock_type, (file - m_file)));
+                        lock_type, (int) (file - m_file)));
     if ((error= (*file)->external_lock(thd, lock_type)))
     {
       if (F_UNLCK != lock_type)
@@ -2508,7 +2508,7 @@ THR_LOCK_DATA **ha_partition::store_lock(THD *thd,
   file= m_file;
   do
   {
-    DBUG_PRINT("info", ("store lock %d iteration", (file - m_file)));
+    DBUG_PRINT("info", ("store lock %d iteration", (int) (file - m_file)));
     to= (*file)->store_lock(thd, to, lock_type);
   } while (*(++file));
   DBUG_RETURN(to);
